Preliminary cost calculation
Corrected screen X positioning calculation.

Added function to calculate cost as options are changed and update the footer with the value. Currently using set data, production version should query server for this info (which should simplify the javascript).

Footer now stays visible at the bottom of the page. Overflow set on form content to give appearance of the questions scrolling beneath it.