Restricted choice dialog, code cleanup
Swapped Y and Z in HTML to match Three.js axis orientation.

Instead of storing references to elements and then getting their values when their functions are called, now when listeners call their functions the values of that element is passed directly into the function. This seems like it will be more efficient memory-wise, and simplifies the code.

Added dialog popup that shows when you choose a restricted option. Gives option of sending an email or resetting that choice.