More form options, now uses basic Python script for calculations
Added options in the form to set wall material and choose a writing surface (whiteboard, blackboard, etc.). Changed lectern type to equipment type as this will be used to choose the default configurations when this has been decided.

Re-ordered the options in the form for clearer user flow and updated code accordingly: for clarity the various loops and functions follow the same order as the form.

Completed basic PHP/Python implementation. Variables are passed to PHP which then calls the Python script. This reads the variables and processes the data before passing them back as a JSON-formatted string. This is returned to the JavaScript which parses the JSON and processes the data as needed.